Use a real args parser
