+++ tests/options-test.mts
@@ -16,3 +16,4 @@ describe('resloveOptions', () => {
     });
   });
 });
+
